+++ submodules/TelegramUI/Sources/Chat/ChatControllerOpenPeer.swift
@@ -401,7 +401,7 @@ extension ChatControllerImpl {
             self?.beginMessageSearch("")
         })))
         if self.context.sharedContext.immediateForkExtrasSettings.saveDeletedMessages {
-            items.append(.action(ContextMenuActionItem(text: "View Deleted", icon: { theme in
+            items.append(.action(ContextMenuActionItem(text: ForkMessageSavingStrings.viewDeleted, icon: { theme in
                 return generateTintedImage(image: UIImage(bundleImageName: "Chat/Context Menu/Archive"), color: theme.contextMenu.primaryColor)
             }, action: { [weak self] action in
                 action.dismissWithResult(.default)